+++ R/db-sql.R
@@ -101,12 +101,12 @@ sql_expr_matches <- function(con, x, y, ...) {
 # https://modern-sql.com/feature/is-distinct-from
 #' @export
 sql_expr_matches.DBIConnection <- function(con, x, y, ...) {
-  build_sql(
-    "CASE WHEN (", x, " = ", y, ") OR (", x, " IS NULL AND ", y, " IS NULL) ",
+  glue_sql2(
+    con,
+    "CASE WHEN ({x} = {y}) OR ({x} IS NULL AND {y} IS NULL) ",
     "THEN 0 ",
     "ELSE 1 ",
-    "END = 0",
-    con = con
+    "END = 0"
   )
 }
 
@@ -147,7 +147,7 @@ sql_table_analyze <- function(con, table, ...) {
 }
 #' @export
 sql_table_analyze.DBIConnection <- function(con, table, ...) {
-  build_sql("ANALYZE ", as.sql(table, con = con), con = con)
+  glue_sql2(con, "ANALYZE {.tbl {table}}")
 }
 
 #' @rdname db-sql
@@ -175,11 +175,10 @@ sql_table_index.DBIConnection <- function(con,
                                           ...,
                                           call = caller_env()) {
   name <- name %||% paste0(c(unclass(table), columns), collapse = "_")
-  fields <- escape(ident(columns), parens = TRUE, con = con)
-  build_sql(
-    "CREATE ", if (unique) sql("UNIQUE "), "INDEX ", as.sql(name, con = con),
-    " ON ", as.sql(table, con = con), " ", fields,
-    con = con
+  glue_sql2(
+    con,
+    "CREATE ", if (unique) "UNIQUE ", "INDEX {.name name}",
+    " ON {.tbl table} ({.col columns*})"
   )
 }
 
@@ -194,7 +193,7 @@ sql_query_explain <- function(con, sql, ...) {
 }
 #' @export
 sql_query_explain.DBIConnection <- function(con, sql, ...) {
-  build_sql("EXPLAIN ", sql, con = con)
+  glue_sql2(con, "EXPLAIN {sql}")
 }
 
 #' @rdname db-sql
@@ -222,10 +221,11 @@ sql_query_save <- function(con, sql, name, temporary = TRUE, ...) {
 }
 #' @export
 sql_query_save.DBIConnection <- function(con, sql, name, temporary = TRUE, ...) {
-  build_sql(
+  glue_sql2(
+    con,
     "CREATE ", if (temporary) sql("TEMPORARY "), "TABLE \n",
-    as.sql(name, con), " AS\n", sql,
-    con = con
+    "{.tbl {name}} AS\n",
+    sql
   )
 }
 #' @export
@@ -238,13 +238,30 @@ sql_query_wrap <- function(con, from, name = NULL, ..., lvl = 0) {
 }
 #' @export
 sql_query_wrap.DBIConnection <- function(con, from, name = NULL, ..., lvl = 0) {
+  sql_query_wrap_helper(
+    con = con,
+    from = from,
+    name = name,
+    lvl = lvl,
+    as = FALSE
+  )
+}
+
+sql_query_wrap_helper <- function(con, from, name, ..., lvl, as) {
   if (is.ident(from)) {
-    setNames(from, name)
-  } else if (is_schema(from) || is_catalog(from)) {
-    setNames(as.sql(from, con), name)
-  } else {
-    build_sql(sql_indent_subquery(from, con, lvl), " ", as_subquery_name(name), con = con)
+    out <- setNames(from, name)
+    return(out)
   }
+
+  if (is_schema(from) || is_catalog(from) || inherits(from, "Id")) {
+    out <- setNames(as.sql(from, con), name)
+    return(out)
+  }
+
+  from <- sql_indent_subquery(from, con, lvl)
+  # some backends, e.g. Postgres, require an alias for a subquery
+  name <- as_subquery_name(name)
+  glue_sql2(con, "{.sql from}", if (as) " AS", " {.name name}")
 }
 
 as_subquery_name <- function(x, default = ident(unique_subquery_name())) {
@@ -286,7 +303,7 @@ sql_query_rows <- function(con, sql, ...) {
 #' @export
 sql_query_rows.DBIConnection <- function(con, sql, ...) {
   from <- dbplyr_sql_subquery(con, sql, "master")
-  build_sql("SELECT COUNT(*) FROM ", from, con = con)
+  glue_sql2(con, "SELECT COUNT(*) FROM {.from from}")
 }
 
 #' @rdname db-sql
@@ -581,7 +598,7 @@ sql_query_semi_join.DBIConnection <- function(con,
   lines <- list(
     sql_clause_select(con, vars),
     sql_clause_from(x),
-    build_sql("WHERE ", if (anti) sql("NOT "), "EXISTS (", con = con),
+    glue_sql2(con, "WHERE ", if (anti) "NOT ", "EXISTS ("),
     # lvl = 1 because they are basically in a subquery
     sql_clause("SELECT 1 FROM", y, lvl = 1),
     # don't use `sql_clause_where()` to avoid wrapping each element in parens